Adds pattern synonym for Vector.
As it turns out, pattern matching on the two coordinates that constitute
a Vector is quite the frequent operation, and having to jump through the
Point constructor each time is very annoying. Therefore, the Vector
'constructor' is now a pattern synonym, permitting immediate access to
the coordinates (both for pattern matching and construction).

Annoyingly, pattern exhaustiveness checking does not seem to be properly
implemented in GHC yet, so using the synonym leads to a lot of warnings
with -Wall. As a temporary workaround, we disable warnings for
non-exhaustive patterns globally, which is certainly less than ideal.