ex06_1: Implements history size limiting
This turned out to be a major problem in many Java submissions for this
exercise, so I've modified the Haskell solution to more accurately
reflect the exercise's difficulty.

The size of a spreadsheet history can now be limited. If a history
grows bigger than the size limit, it drops the earliest stored actions
until it is within the size limit again.

Internally, the data structure for the history is now a
Data.Sequence.Seq rather than a list because it has to support efficient
deletions from the end.